jiaendu restrict length

+++ benchmarks/0000.10m_size_t/fast_io_vs_fmt.cc
@@ -3,13 +3,13 @@
 
 int main()
 {
-	constexpr std::size_t N(10000000);
+	constexpr std::uint32_t N(10000000);
 	std::size_t fast_io_ostr_value{};
 	{
 	fast_io::timer t("fast_io::ostring");
 	fast_io::ostring ostr;
 	ostr.str().reserve(42);
-	for(std::size_t i{};i!=N;++i)
+	for(std::uint32_t i{};i!=N;++i)
 	{
 		print(ostr,i);
 		fast_io_ostr_value+=ostr.str().size();
@@ -19,13 +19,13 @@ int main()
 	std::size_t fast_io_value{};
 	{
 	fast_io::timer t("fast_io::concat");
-	for(std::size_t i{};i!=N;++i)
+	for(std::uint32_t i{};i!=N;++i)
 		fast_io_value+=fast_io::concat(i).size();
 	}
 	std::size_t fmt_value{};
 	{
 	fast_io::timer t("fmt::format");
-	for(std::size_t i{};i!=N;++i)
+	for(std::uint32_t i{};i!=N;++i)
 		fmt_value+=fmt::format_int(i).str().size();
 	}
 	println("fast_io::ostring value:",fast_io_ostr_value,"\nfast_io::concat value:",fast_io_value,"\nfmt::format value:",fmt_value);
